apps/ppe-detection/requirement.txt
cvzone
fastapi
ultralytics
uvicorn
python-multipart
numpy<2.0
pydantic-settings